fix(setup): reconcile guidance + register MCP after hosts install
On a genuinely bare machine, machine setup registered the user-scope MCP
server and synced the CLAUDE.md guidance blocks before the hosts branch had
installed the claude/codex/opencode CLIs — so first runs warned ('claude mcp
add failed') and reported immediate block drift (command:codex and
flag:dualMode detectors flip only after the hosts step). Both steps now run
as a final pass in the setup orchestrator, after the hosts branch and the
project phase's codex bridge (which creates ~/.codex). The guidance
reconcile loop is extracted to blocks.mjs reconcileGuidance and shared with
ak sync, so the two commands converge guidance identically.
